add form

+++ STO/BuisnessLogic/BindingModels/PaymentBindingModel.cs
@@ -3,8 +3,8 @@
     public class PaymentBindingModel
     {
         public int? Id { get; set; }
-        public int Sum { get; set; }
-        public int Remain { get; set; }
+        public decimal Sum { get; set; }
+        public decimal Remain { get; set; }
         public int RepairId { get; set; }
     }
 }
