+++ STO/BuisnessLogic/BindingModels/WorkBindingModel.cs
@@ -10,7 +10,7 @@ namespace BuisnessLogic.BindingModels
     {
         public int? Id { get; set; }
         public string WorkName { get; set; }
-        public int WorkPrice { get; set; }
+        public decimal WorkPrice { get; set; }
 
     }
 }
