+++ STO/BuisnessLogic/ViewModels/PaymentViewModel.cs
@@ -3,9 +3,9 @@
     public class PaymentViewModel
     {
         public int Id { get; set; }
-        public int Remain { get; set; }
+        public decimal Remain { get; set; }
         public int RepairId { get; set; }
-        public int Sum { get; set; }
+        public decimal Sum { get; set; }
         public string RepairName { get; set; }
     }
 }
